Refactor layout and fix context imports
