+++ components/Layout.js
@@ -1,10 +1,9 @@
 import React from "react";
-import { CartProvider } from "../contexts/CartContext";
 import Link from "next/link";
 
 const Layout = ({ children }) => {
   return (
-    <CartProvider>
+    <>
       <header className="bg-gray-800 text-white p-4">
         <div className="container mx-auto flex justify-between items-center">
           <Link href="/" className="text-lg font-bold">Hopeflex</Link>
@@ -19,7 +18,7 @@ const Layout = ({ children }) => {
       <footer className="bg-gray-800 text-white p-4 text-center">
         <p>© 2025 Hopeflex Store</p>
       </footer>
-    </CartProvider>
+    </>
   );
 };
