+++ pages/account.js
@@ -1,11 +1,10 @@
-import Layout from '../components/Layout';
 
 export default function Account() {
   // In a real app, fetch user orders from DB
   return (
-    <Layout>
+      <>
       <h1>Meus Pedidos</h1>
       <p>Funcionalidade em desenvolvimento.</p>
-    </Layout>
+    </>
   );
 }
