+++ pages/admin/panel.js
@@ -1,9 +1,8 @@
 import Link from 'next/link';
-import Layout from '../../components/Layout';
 
 export default function Admin() {
   return (
-    <Layout>
+      <>
       <h1>Adicionar Produto</h1>
       <form action="/api/admin" method="POST">
         <input type="text" name="name" placeholder="Nome" required /><br />
@@ -12,6 +11,6 @@ export default function Admin() {
         <button type="submit">Criar</button>
       </form>
       <Link href="/">Voltar</Link>
-    </Layout>
+    </>
   );
 }
